refactor: enhance ArtistAccessSelector by integrating HasAccessList component
- Replaced inline access display logic in ArtistAccessSelector with the new HasAccessList component for improved code organization and readability.
- Updated the useFileAccessGrant hook to fetch and manage access data, streamlining the UI loading state handling.
- Introduced shared types for artist access management in types.ts to maintain consistency across components.
