+++ components/Files/types.ts
@@ -7,3 +7,19 @@ export interface FileRow {
 }
 
 
+// Shared types for File Access UI
+export type ArtistAccess = {
+  artistId: string;
+  scope: "read_only" | "admin";
+  grantedAt: string;
+  expiresAt: string | null;
+  artistName?: string | null;
+  artistEmail?: string | null;
+};
+
+export type AccessArtistsResponse = {
+  success: boolean;
+  data?: { artists: ArtistAccess[]; count: number; fileId: string; accountId: string };
+};
+
+
