+++ hooks/useFileAccessGrant.tsx
@@ -1,6 +1,8 @@
 import { useCallback, useState } from "react";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQueryClient, useQuery } from "@tanstack/react-query";
 import { toast } from "sonner";
+import useUser from "@/hooks/useUser";
+import type { ArtistAccess, AccessArtistsResponse } from "@/components/Files/types";
 
 type AccessGrant = {
   fileId: string;
@@ -11,6 +13,8 @@ type AccessGrant = {
 const useFileAccessGrant = ({ fileId, grantedBy, invalidateKeys }: AccessGrant) => {
   const [selected, setSelected] = useState<string[]>([]);
   const queryClient = useQueryClient();
+  const { userData } = useUser();
+  const accountId = userData?.account_id;
   const { mutate, isPending: isSavingAccess } = useMutation({
     mutationFn: async (selected: string[]) => {
       const response = await fetch("/api/files/grant-access", {
@@ -49,6 +53,23 @@ const useFileAccessGrant = ({ fileId, grantedBy, invalidateKeys }: AccessGrant) 
     },
   });
 
+  // Fetch existing access data
+  const { data: accessData, isLoading: isAccessLoading } = useQuery<AccessArtistsResponse | null>({
+    queryKey: ["file-access-artists", fileId, accountId],
+    queryFn: async () => {
+      if (!accountId) return null;
+      const res = await fetch(`/api/files/access-artists?fileId=${encodeURIComponent(fileId)}&accountId=${encodeURIComponent(accountId)}`);
+      if (!res.ok) return null;
+      return res.json();
+    },
+    enabled: Boolean(fileId && accountId),
+    staleTime: 30_000,
+  });
+
+  // Treat UI as loading until we have an accountId and the query has resolved
+  const isUiLoading = !accountId || isAccessLoading;
+  const alreadyHasAccessIds: string[] = (accessData?.data?.artists || []).map((a: ArtistAccess) => a.artistId);
+
   const add = useCallback(
     (id: string) =>
       setSelected((prev) => (prev.includes(id) ? prev : [...prev, id])),
@@ -71,6 +92,9 @@ const useFileAccessGrant = ({ fileId, grantedBy, invalidateKeys }: AccessGrant) 
     remove,
     onGrant,
     isSavingAccess,
+    accessData,
+    isUiLoading,
+    alreadyHasAccessIds,
   };
 };
